% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{fread_solani}
\alias{fread_solani}
\title{Read Solani File using data.table}
\usage{
fread_solani(ps_path)
}
\arguments{
\item{ps_path}{path to input file}
}
\value{
DataTable with content of solani file
}
\description{
The solani result file from MiX99 is read using `data.table::fread()`
}
\examples{
s_ebv_path_partial <- qzwslrm_example_solani("partial")
dt_solani <- fread_solani(ps_path = s_ebv_path_partial)
}
