% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qzwslrm_logger.R
\name{get_qzwslrm_logger}
\alias{get_qzwslrm_logger}
\title{Create log4r Logger for package}
\usage{
get_qzwslrm_logger(ps_logfile = "qzwslrm.log", ps_level = "FATAL")
}
\arguments{
\item{ps_logfile}{name of the logfile}

\item{ps_level}{logger level}
}
\value{
qzwslrm_logger
}
\description{
Create log4r Logger for package
}
\examples{
\dontrun{
qzwslrm_logger <- get_qzwslrm_logger()
}
}
