% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{qzwslrm_example}
\alias{qzwslrm_example}
\title{Return Path to Example Data}
\usage{
qzwslrm_example(ps_data_type)
}
\arguments{
\item{ps_data_type}{type of data set can be either 'partial' or 'whole'}
}
\value{
path to example data
}
\description{
Return path to example data
}
\examples{
ed_path <- qzwslrm_example('partial')
}
