% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{qzwslrm_example_input}
\alias{qzwslrm_example_input}
\title{Path to Input Example Files}
\usage{
qzwslrm_example_input(ps_data_format)
}
\arguments{
\item{ps_data_format}{format of input file}
}
\value{
path to input file
}
\description{
The path to the input files with different formats contained in this package
can be obtained
}
\examples{
sexdf_csv <- qzwslrm_example_input("csv")
}
