% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{qzwslrm_example_solani}
\alias{qzwslrm_example_solani}
\title{Return Path to Example Data}
\usage{
qzwslrm_example_solani(ps_data_type)
}
\arguments{
\item{ps_data_type}{type of data set can be either 'partial' or 'whole'}
}
\value{
path to example data
}
\description{
Return path to example solani result file containing EBV from either whole
or partial data.
}
\examples{
ed_path <- qzwslrm_example_solani('partial')
}
