% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qzwslrm_logger.R
\name{qzwslrm_log_info}
\alias{qzwslrm_log_info}
\title{Wrapper for log4r info}
\usage{
qzwslrm_log_info(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_caller}{function from which we are called}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r info
}
\examples{
\dontrun{
qzwslrm_logger <- get_qzwslrm_logger()
qzwslrm_log_level(qzwslrm_logger, 'INFO')
qzwslrm_log_info(qzwslrm_logger, 'Examples', 'test message')
}
}
