% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qzwslrm_logger.R
\name{qzwslrm_log_level}
\alias{qzwslrm_log_level}
\title{Wrapper to set the level of a logger}
\usage{
qzwslrm_log_level(
  plogger,
  ps_level = c("DEBUG", "INFO", "WARN", "ERROR", "FATAL")
)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_level}{new level of plogger}
}
\description{
Wrapper to set the level of a logger
}
\examples{
\dontrun{
qzwslrm_logger <- get_qzwslrm_logger()
qzwslrm_log_level(qzwslrm_logger, 'INFO')
}
}
