% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read.solani}
\alias{read.solani}
\title{Read Solani File using R-Base read.table}
\usage{
read.solani(ps_path)
}
\arguments{
\item{ps_path}{path to input file}
}
\value{
dataframe with content of solani file
}
\description{
The function `read.table()` from R-base is used to read the solani file.
}
\examples{
s_ebv_path_partial <- qzwslrm_example_solani("partial")
df_solani <- read.solani(ps_path = s_ebv_path_partial)
}
