% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{readr_ebv}
\alias{readr_ebv}
\title{Read Estimated Breeding Values (EBV) from Input File}
\usage{
readr_ebv(
  ps_path,
  ps_format,
  ps_delim_char = " ",
  ps_animal_col_name = NULL,
  pn_animal_col_idx = 1,
  ps_ebv_col_name = NULL,
  pn_ebv_col_idx = NULL
)
}
\arguments{
\item{ps_path}{path to the EBV input file}

\item{ps_format}{format of the EBV input file, can be either of "csv", "csv2", "delim" or "table"}

\item{ps_delim_char}{column delimiting character for format "delim"}

\item{ps_animal_col_name}{column name for animal ID column}

\item{pn_animal_col_idx}{column index for animal ID column}

\item{ps_ebv_col_name}{column name for ebv column}

\item{pn_ebv_col_idx}{column index for ebv column}
}
\value{
ebv result tibble
}
\description{
Given a result file containing EBV and possibly other information,
the column with AnimalID and with EBV are extracted and returned as
a two column result tibble. The result tibble contains the AnimalID
in the first column and the EBV in the second column.
}
\details{
The EBV input file can be in different formats. Possible formats that
are considered can be "csv", "csv2", "delim", "table". Where each of
these format specifiers point to the respective function from the `readr`
package that is used for importing.
}
\examples{
s_ebv_path_partial <- qzwslrm_example_solani("partial")
tbl_ebv <- readr_ebv(ps_path = s_ebv_path_partial, ps_format = "table")
}
