% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{readr_solani}
\alias{readr_solani}
\title{Read Solani File Using readr}
\usage{
readr_solani(
  ps_path,
  pvec_col_names = c("animal", "nrprg", "trait", "ebv"),
  pl_col_types = readr::cols(animal = readr::col_integer(), nprg =
    readr::col_integer(), trait = readr::col_integer(), ebv = readr::col_double())
)
}
\arguments{
\item{ps_path}{path to input file}

\item{pvec_col_names}{vector of column names}

\item{pl_col_types}{col type specifications using `readr::cols()`}
}
\value{
tibble with predicted breeding values
}
\description{
The file Solani contains predicted breeding values computed by
MiX99. This result file is read using `readr::read_table()`
}
\examples{
s_ebv_path_partial <- qzwslrm_example_solani("partial")
tbl_solani_partial <- readr_solani(ps_path = s_ebv_path_partial)
}
