% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scatterplot_lrm}
\alias{scatterplot_lrm}
\title{Scatter Plot Based on Validation Input}
\usage{
scatterplot_lrm(ptbl_ebv_whole, ptbl_ebv_partial)
}
\arguments{
\item{ptbl_ebv_whole}{ebv based on whole data}

\item{ptbl_ebv_partial}{ebv based on partial data}
}
\value{
ggplot2 objection conaining scatterplot
}
\description{
The EBV for which validation is done is shown in a scatterplot
}
\examples{
tbl_ebv_whole <- readr_ebv(ps_path = qzwslrm_example_solani("whole"), ps_format = "table",
                          pn_ebv_col_idx = 4)
tbl_ebv_partial <- readr_ebv(ps_path = qzwslrm_example_solani("partial"), ps_format = "table",
                            pn_ebv_col_idx = 4)
p <- scatterplot_lrm(tbl_ebv_whole, tbl_ebv_partial)
}
