% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{summary_lrm}
\alias{summary_lrm}
\title{Summary of EBV Validation Results}
\usage{
summary_lrm(object, pn_digits = max(3L, getOption("digits") - 3L))
}
\arguments{
\item{object}{validation result}

\item{pn_digits}{number of digits to be used for rounding the results}
}
\description{
The list with the validation results is printed in a summary format.
}
\examples{
tbl_ebv_whole <- readr_ebv(ps_path = qzwslrm_example_solani("whole"), ps_format = "table",
                           pn_ebv_col_idx = 4)
tbl_ebv_partial <- readr_ebv(ps_path = qzwslrm_example_solani("partial"), ps_format = "table",
                             pn_ebv_col_idx = 4)
summary_lrm(l_val_result <- val_ebv_lrm(pvec_ebv_partial = tbl_ebv_partial$ebv,
                                        pvec_ebv_whole = tbl_ebv_whole$ebv))
}
