% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_ebv.R
\name{val_ebv_lrm}
\alias{val_ebv_lrm}
\title{Validation of Estimated Breeding Values (EBV) Using LR-Method}
\usage{
val_ebv_lrm(pvec_ebv_partial, pvec_ebv_whole)
}
\arguments{
\item{pvec_ebv_partial}{vector of EBV from partial data}

\item{pvec_ebv_whole}{vector of EBV from whole data}
}
\value{
list of results containing bias, correlation and regression coefficients
}
\description{
Validation statistics for EBV based on bias, correlation and regression
coefficients between EBV results based on partial and whole data sets are
computed.
}
\examples{

}
